+++ edgy/contrib/admin/templates/admin/base.html
@@ -6,7 +6,7 @@
     <link rel="icon" href="{{ favicon }}"/>
     <title>{% block title %}Edgy Admin | {{ title }}{% endblock %}</title>
     <script src="https://cdn.tailwindcss.com"></script>
-    <link rel="stylesheet" type="text/css" href="https://unpkg.com/@fortawesome/fontawesome-free@6.7.2/css/fontawesome.min.css"></link>
+    <link rel="stylesheet" type="text/css" href="https://unpkg.com/@fortawesome/fontawesome-free@6.7.2/css/all.min.css"></link>
     <script src="https://cdn.jsdelivr.net/npm/@json-editor/json-editor@latest/dist/jsoneditor.min.js"></script>
     <script>
         JSONEditor.defaults.theme = 'tailwind';
