+++ edgy/contrib/admin/templates/admin/model_detail.html
@@ -39,17 +39,12 @@
     </a>
 </div>
 
-{% if objects %}
+{% if page.content %}
 <div class="overflow-x-auto bg-white rounded-lg shadow border">
     <table class="min-w-full divide-y divide-gray-200">
         <thead class="bg-gray-50">
             <tr>
-                {% if 'id' in model.__fields__.keys() %}
-                <th scope="col" class="px-6 py-3 text-left text-xs font-semibold text-gray-600 uppercase tracking-wider">
-                    id
-                </th>
-                {% endif %}
-                {% for field in model.__fields__.keys() if field != 'id' %}
+                {% for field in model.__fields__.keys() if field != 'pk' %}
                 <th scope="col" class="px-6 py-3 text-left text-xs font-semibold text-gray-600 uppercase tracking-wider">
                     {{ field.replace('_', ' ') | title }}
                 </th>
@@ -57,14 +52,9 @@
             </tr>
         </thead>
         <tbody class="divide-y divide-gray-100 bg-white">
-            {% for obj in objects %}
-            <tr class="hover:bg-gray-50 transition cursor-pointer" onclick="window.location='{{ url_prefix }}/models/{{ model_name }}/{{ obj.id }}'">
-                {% if 'id' in model.__fields__.keys() %}
-                <td class="px-6 py-4 text-sm text-gray-800 whitespace-nowrap">
-                    {{ getattr(obj, 'id') }}
-                </td>
-                {% endif %}
-                {% for field in model.__fields__.keys() if field != 'id' %}
+            {% for obj in page.content %}
+            <tr class="hover:bg-gray-50 transition cursor-pointer" onclick="window.location='{{ url_prefix }}/models/{{ model_name }}/{{ create_object_pk(obj.pk) }}'">
+                {% for field in model.__fields__.keys() if field != 'pk' %}
                 <td class="px-6 py-4 text-sm text-gray-800 whitespace-nowrap overflow-hidden text-ellipsis max-w-xs">
                     {{ getattr(obj, field) }}
                 </td>
@@ -82,8 +72,8 @@
 
 {% if page and total_pages %}
 <div class="mt-6 flex justify-center space-x-2 text-sm">
-    {% if page > 1 %}
-    <a href="?page={{ page - 1 }}&per_page={{ per_page }}{% if query %}&q={{ query }}{% endif %}"
+    {% if not page.is_first %}
+    <a href="?page={{ page.previous_page }}&per_page={{ per_page }}{% if query %}&q={{ query }}{% endif %}"
        class="px-3 py-1 bg-gray-200 text-gray-700 rounded hover:bg-gray-300">Previous</a>
     {% endif %}
 
@@ -92,8 +82,8 @@
        class="px-3 py-1 {{ 'bg-blue-600 text-white' if p == page else 'bg-gray-100 text-gray-800' }} rounded hover:bg-gray-200">{{ p }}</a>
     {% endfor %}
 
-    {% if page < total_pages %}
-    <a href="?page={{ page + 1 }}&per_page={{ per_page }}{% if query %}&q={{ query }}{% endif %}"
+    {% if not page.is_last %}
+    <a href="?page={{ page.next_page }}&per_page={{ per_page }}{% if query %}&q={{ query }}{% endif %}"
        class="px-3 py-1 bg-gray-200 text-gray-700 rounded hover:bg-gray-300">Next</a>
     {% endif %}
 </div>
