+++ edgy/contrib/admin/templates/admin/model_object_create.html
@@ -7,8 +7,8 @@
 {% block content %}
 <div id="editor_holder"></div>
 <div class="pt-6 flex gap-4">
-  <form>
-    <button
+  <form method="post" id="submit_form">
+    <button type="submit"
             class="px-5 py-2 bg-green-600 text-white text-sm font-medium rounded hover:bg-green-700">
       Create
     </button>
@@ -23,12 +23,12 @@
 <script>
 
       var editor = new JSONEditor(document.getElementById('editor_holder'),{
-        // Enable fetching schemas via ajax
         ajax: true,
+        ajax_cache_responses: true,
         required_by_default: false,
         show_opt_in: true,
-        ajax_cache_responses: true,
-        disable_edit_json: true,
+        collapsed: true,
+        disable_properties: true,
 
         // The schema for the editor
         schema: {
@@ -37,9 +37,10 @@
         },
       });
 
-      const editor_data = document.getElementById("editor_data")
-      editor.on('change',() => {
+      var editor_data = document.getElementById("editor_data")
+      document.getElementById("submit_form").addEventListener("submit", async function(ev){
         editor_data.value = JSON.stringify(editor.getValue())
-      });
+
+      })
 </script>
 {% endblock %}
