+++ edgy/contrib/admin/templates/admin/model_object_edit.html
@@ -1,141 +1,42 @@
 {% extends "admin/base.html" %}
 
-{% block header %}Edit {{ model_name | replace('_', ' ') | title }} #{{ object.id }}{% endblock %}
+{% block header %}Edit {{ model_name | replace('_', ' ') | title }} #{{ object }}{% endblock %}
 
 {% block content %}
-<form method="post" class="bg-white rounded-lg shadow p-6 min-w-full mx-auto space-y-6">
-
-    {# Regular fields (non-relationship) #}
-    {% for field, field_info in model.model_fields.items() %}
-        {% if field != "pk" and not relationship_fields.get(field) %}
-        <div class="flex flex-col gap-1">
-            <label class="text-sm font-semibold text-gray-700" for="{{ field }}">
-                {{ field.replace('_', ' ') | title }}
-                {% if not field_info.null %}
-                  <span class="text-red-500 ml-1" title="Required">*</span>
-                {% endif %}
-            </label>
-
-            {% set value = getattr(object, field) %}
-            {% set type = get_input_type(field_info.annotation) %}
-
-            {% if type == "bool" %}
-            <select name="{{ field }}" id="{{ field }}" class="border border-gray-300 rounded px-3 py-2 text-sm text-gray-900" {% if field_info.read_only %}disabled{% endif %}>
-                <option value="1" {% if value is sameas true %}selected{% endif %}>True</option>
-                <option value="0" {% if value is sameas false %}selected{% endif %}>False</option>
-            </select>
-
-            {% elif type == "date" %}
-            <input
-                type="date"
-                name="{{ field }}"
-                id="{{ field }}"
-                value="{{ value }}"
-                class="border border-gray-300 rounded px-3 py-2 text-sm text-gray-900 focus:ring focus:ring-blue-200"
-                {% if field_info.read_only %}disabled{% endif %}
-            />
-
-            {% elif type == "datetime" %}
-            <input
-                type="datetime-local"
-                name="{{ field }}"
-                id="{{ field }}"
-                value="{{ value.isoformat() if value else '' }}"
-                class="border border-gray-300 rounded px-3 py-2 text-sm text-gray-900 focus:ring focus:ring-blue-200"
-                {% if field_info.read_only %}disabled{% endif %}
-            />
-
-            {% else %}
-            <input
-                type="text"
-                name="{{ field }}"
-                id="{{ field }}"
-                value="{{ value }}"
-                class="border border-gray-300 rounded px-3 py-2 text-sm text-gray-900 focus:ring focus:ring-blue-200"
-                {% if field_info.read_only %}disabled{% endif %}
-            />
-            {% endif %}
-        </div>
-        {% endif %}
-    {% endfor %}
-
-    {# Relationship fields (FK and M2M) #}
-    {% for field, rel in relationship_fields.items() %}
-    <div class="flex flex-col gap-1">
-        <label class="text-sm font-semibold text-gray-700" for="{{ field }}">
-            {{ field.replace('_', ' ') | title }}
-        </label>
-
-        {% if rel["type"] == "foreign_key" %}
-        <select name="{{ field }}" id="{{ field }}" class="border border-gray-300 rounded px-3 py-2 text-sm text-gray-900" {% if field_info.read_only %}disabled{% endif %}>
-            <option value="">Select...</option>
-            {% for obj in rel["items"] %}
-                <option value="{{ obj.id }}" {% if getattr(object, field).id == obj.id %}selected{% endif %}>{{ obj }}</option>
-            {% endfor %}
-        </select>
-
-        {% elif rel["type"] == "many_to_many" %}
-        {% set selected = rel["selected"] %}
-        <input type="hidden" name="_{{ field }}_initial" value="{{ selected | join(',') }}">
-
-        <div class="flex gap-4 items-center">
-            <select id="{{ field }}_available" class="w-full border rounded h-48 text-sm" multiple {% if field_info.read_only %}disabled{% endif %}>
-                {% for obj in rel["items"] %}
-                    {% if obj.id not in selected %}
-                    <option value="{{ obj.id }}">{{ obj }}</option>
-                    {% endif %}
-                {% endfor %}
-            </select>
-
-            <div class="flex flex-col gap-2">
-                <button type="button" onclick="moveSelected('{{ field }}_available', '{{ field }}_selected')" class="text-sm bg-gray-300 rounded px-2 py-1 hover:bg-gray-400">→</button>
-                <button type="button" onclick="moveSelected('{{ field }}_selected', '{{ field }}_available')" class="text-sm bg-gray-300 rounded px-2 py-1 hover:bg-gray-400">←</button>
-            </div>
-
-            <select id="{{ field }}_selected" name="{{ field }}" multiple class="w-full border rounded h-48 text-sm" {% if field_info.read_only %}disabled{% endif %}>
-                {% for obj in rel["items"] %}
-                    {% if obj.id in selected %}
-                    <option value="{{ obj.id }}">{{ obj }}</option>
-                    {% endif %}
-                {% endfor %}
-            </select>
-        </div>
-        {% endif %}
-    </div>
-    {% endfor %}
-
-    <div class="pt-6 flex gap-4">
+<div id="editor_holder"></div>
+<div class="pt-6 flex gap-4">
+  <form method="post" id="submit_form">
         <button type="submit" class="px-5 py-2 bg-blue-600 text-white text-sm font-medium rounded hover:bg-blue-700">
             Save Changes
         </button>
-        <a href="{{ url_prefix }}/models/{{ model_name }}/{{ object.id }}" class="px-5 py-2 bg-gray-200 text-gray-800 text-sm font-medium rounded hover:bg-gray-300">
-            Cancel
-        </a>
-    </div>
-</form>
-
+        <input type="hidden" name="editor_data" id="editor_data" value=''>
+    </form>
+    <a href=".." class="px-5 py-2 bg-gray-200 text-gray-800 text-sm font-medium rounded hover:bg-gray-300">
+        Cancel
+    </a>
+</div>
 <script>
 
-    var editor = new JSONEditor(document.getElementById('editor_holder'),{
-    // Enable fetching schemas via ajax
-    ajax: true,
+      var editor = new JSONEditor(document.getElementById('editor_holder'),{
+        ajax: true,
+        ajax_cache_responses: true,
+        required_by_default: false,
+        show_opt_in: true,
+        collapsed: true,
+        disable_properties: true,
 
-    // The schema for the editor
-    schema: {
-        $ref: "person.json",
-        format: "grid"
-    },
+        // The schema for the editor
+        schema: {
+          $ref: "{{ url_prefix }}/models/{{ model_name }}/json?cdefaults=true",
+          format: "grid",
+        },
+        startval: JSON.parse('{{object.model_dump_json() | safe}}')
+      });
 
-    // Seed the form with a starting value
-    startval: starting_value
-    });
+      const editor_data = document.getElementById("editor_data")
+      document.getElementById("submit_form").addEventListener("submit", async function(ev){
+        editor_data.value = JSON.stringify(editor.getValue())
 
-    // Hook up the submit button to log to the console
-    document.getElementById('submit').addEventListener('click',function() {
-    // Get the value from the editor
-    console.log(editor.getValue());
-    });
+      })
 </script>
-
-
 {% endblock %}
