--- edgy/contrib/admin/templates/admin/model_object_view.html
@@ -1,63 +1,0 @@
-{% extends "admin/base.html" %}
-
-{% block header %}{{ model_name | replace('_', ' ') | title }} #{{ object.id }}{% endblock %}
-
-{% block content %}
-<div class="bg-white border border-gray-200 rounded-lg shadow-sm w-full px-6 py-6">
-
-    <div class="space-y-6">
-        {% for field, field_info in model.__fields__.items() %}
-        <div class="flex flex-col sm:flex-row sm:items-start sm:justify-between gap-2">
-            <dt class="text-sm font-semibold text-gray-600 w-full sm:w-1/4">
-                {{ field.replace('_', ' ') | title }}
-            </dt>
-            <dd class="text-sm text-gray-900 bg-gray-50 px-4 py-2 rounded-md border border-gray-100 w-full sm:w-3/4 break-words">
-                {% set value = getattr(object, field) %}
-                {% set rel_type = relationship_fields.get(field) %}
-
-                {% if rel_type == "many_to_many" %}
-                    {% set items = m2m_values.get(field, []) %}
-                    {% if items %}
-                        {% for item in items %}
-                            {{ item }}{% if not loop.last %}, {% endif %}
-                        {% endfor %}
-                    {% else %}
-                        —
-                    {% endif %}
-
-                {% elif rel_type == "foreign_key" %}
-                    {% if value %}
-                        {{ value }}
-                    {% else %}
-                        —
-                    {% endif %}
-
-                {% else %}
-                    {% if value is not none %}
-                        {{ value }}
-                    {% else %}
-                        —
-                    {% endif %}
-                {% endif %}
-            </dd>
-        </div>
-        {% endfor %}
-    </div>
-
-    <div class="mt-8 flex flex-wrap gap-4 border-t pt-6">
-        <a href="{{ url_prefix }}/models/{{ model_name }}/{{ object.id }}/edit"
-           class="inline-block px-5 py-2 bg-blue-600 text-white text-sm font-medium rounded hover:bg-blue-700">
-            Edit
-        </a>
-        <form method="post"
-              action="{{ url_prefix }}/models/{{ model_name }}/{{ object.id }}/delete"
-              onsubmit="return confirm('Are you sure you want to delete this?');">
-            <button type="submit"
-                    class="px-5 py-2 bg-red-600 text-white text-sm font-medium rounded hover:bg-red-700">
-                Delete
-            </button>
-        </form>
-    </div>
-
-</div>
-{% endblock %}
